contact book
This is a basic contact book application. It allows users to store, view, and remove contacts. Each contact consists of a name and a phone number.

Key Features:
Add Contact: The user can add new contacts by entering a name and phone number.
View Contacts: All stored contacts can be viewed in a list with their names and phone numbers.
Remove Contact: The user can remove a contact by entering the name of the contact they wish to delete.
Exit: The app will keep running in a loop until the user chooses to exit.